@@ -25,8 +25,7 @@
  *  AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
  *  ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  **********************************************************************************************************************/
-import {remote} from 'electron';
-const {app} = remote;
+import { app } from '@electron/remote';
 import jetpack from 'fs-jetpack';
 import AdmZip from 'adm-zip';
 import https from 'https';
